backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
httpx>=0.28.0
scenedetect[opencv]>=0.6.7
google-generativeai>=0.8.0
anthropic>=0.49.0
openai>=1.60.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
